Make the Money + Interactables scalars doubles for finer adjustment.

+++ ShareSuite/MoneySharingHooks.cs
@@ -154,7 +154,7 @@ namespace ShareSuite
         private static void GiveAllScaledMoney(float goldReward)
         {
             //Apply gold rewards to shared money pool
-            SharedMoneyValue += (int) Mathf.Floor(goldReward * ShareSuite.MoneyScalar.Value - goldReward);
+            SharedMoneyValue += (int) Mathf.Floor(goldReward * (float) ShareSuite.MoneyScalar.Value - goldReward);
         }
 
         public static void SetTeleporterActive(bool active)
